# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NicoShop API",
    "version": "1.0.0",
    "description": "API REST de l'e-commerce NicoShop"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Serveur local" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "msg": { "type": "string" },
                "path": { "type": "string" },
                "location": { "type": "string" }
              }
            }
          }
        }
      },
...
      "UserPublic": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["customer", "admin"] },
          "picture": { "type": "string" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "token": { "type": "string" },
          "user": { "$ref": "#/components/schemas/UserPublic" }
        }
      },
...
      "Favori": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "product_id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number", "format": "float" },
          "image_url": { "type": "string", "format": "uri" }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Authentification"],
        "summary": "Inscription utilisateur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "Jean Dupont" },
                  "email": { "type": "string", "format": "email", "example": "jean@example.com" },
                  "password": { "type": "string", "minLength": 6, "example": "monmotdepasse" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Inscription réussie", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "400": { "description": "Données invalides ou email déjà utilisé", "content": { "application/json": { "schema": { "oneOf": [ { "$ref": "#/components/schemas/ValidationError" }, { "$ref": "#/components/schemas/Error" } ] } } } },
          "429": { "description": "Trop de tentatives" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentification"],
        "summary": "Connexion utilisateur",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "jean@example.com" },
                  "password": { "type": "string", "example": "monmotdepasse" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Connexion réussie", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "401": { "description": "Email ou mot de passe incorrect" },
          "429": { "description": "Trop de tentatives" }
        }
      }
    },
    "/api/auth/oauth": {
      "post": {
        "tags": ["Authentification"],
        "summary": "Connexion OAuth (Google/Facebook)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "name", "provider", "oauth_id"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "jean@gmail.com" },
                  "name": { "type": "string", "example": "Jean Dupont" },
                  "provider": { "type": "string", "enum": ["google", "facebook"], "example": "google" },
                  "oauth_id": { "type": "string", "example": "123456789" },
                  "picture": { "type": "string", "format": "uri", "example": "https://example.com/photo.jpg" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Connexion OAuth réussie", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "400": { "description": "Données invalides" },
          "429": { "description": "Trop de tentatives" }
        }
      }
    },
...
    "/api/favoris": {
      "get": {
        "tags": ["Favoris"],
        "summary": "Favoris de l'utilisateur connecté",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Liste des favoris", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Favori" } } } } }
        }
      },
      "post": {
        "tags": ["Favoris"],
        "summary": "Ajouter un produit aux favoris",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["product_id"],
                "properties": {
                  "product_id": { "type": "integer", "example": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Favori ajouté", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Favori" } } } },
          "400": { "description": "ID produit requis" },
          "409": { "description": "Déjà en favoris" }
        }
      }
    },
    "/api/favoris/{product_id}": {
      "delete": {
        "tags": ["Favoris"],
        "summary": "Supprimer un favori",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "product_id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }
        ],
        "responses": {
          "200": { "description": "Favori supprimé" },
          "404": { "description": "Favori non trouvé" }
        }
      }
    },
...
  }
}